+++ src/PaperScreening/Presentational/PapersPresentational.js
@@ -1,5 +1,5 @@
 import { PanelGroup } from 'react-bootstrap';
-import PaperPanelPresentational from './PaperPanelPresentational';
+import PaperPanelPresentational from '../Presentational/PaperPanelPresentational';
 import React from 'react';
 
 const PapersPresentational = (props) => {
